reorganized crate
renamed settings.rs to stg.rs and made it a folder

renamed rust_type_implementations.rs to setting_implementations.rs and moved it to the stg folder

made account.rs into a folder

created incrementable_implementations.rs
