--- src/types/mod.rs
@@ -1,2 +1,0 @@
-///module containing implementations of `Setting` for rust types
-pub mod rust_types_implementations;
